Fix clippy lints and touch up GitHub actions scripts.
Bump versions on actions/checkout and fix GPU examples not being tested
properly (running the examples will fail as there's no GPU, but all examples
after the first one were not compiled and thus not checked).

Doing this change helped me find a broken example.

+++ src/analytic/gpu/mod.rs
@@ -67,7 +67,7 @@ impl AnalyticBeamGpu {
             return Err(AnalyticBeamError::IncorrectDelaysArrayColLength {
                 rows: delays_array.len_of(Axis(0)),
                 num_delays: delays_array.len_of(Axis(1)),
-                expected: usize::from(num_bowties),
+                expected: num_bowties,
             });
         }
         if amps_array.len_of(Axis(1)) != num_bowties
